requirements.txt
cycler==0.10.0
h5py==2.7.1
matplotlib==2.1.1
numpy==1.13.3
Pillow==5.0.0
pkg-resources==0.0.0
pygame==1.9.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scipy==1.0.0
six==1.11.0
wget==3.2
